analyze test with database - unit tests
