+++ api/main.py
@@ -1,5 +1,6 @@
 from flask import Flask
 from src.controllers.api import api as api_blueprint
+from src.controllers.analyze_api import api as analyze_api_blueprint
 from src.controllers.env_info_api import env_info_api as env_api_blueprint
 from config.settings import load_config
 from flask_cors import CORS
@@ -12,6 +13,7 @@ load_config(app)
 app.database_manager = DatabaseManager(app)
 # init routes
 app.register_blueprint(api_blueprint, url_prefix='/qujata-api', name='api')
+app.register_blueprint(analyze_api_blueprint, url_prefix='/qujata-api', name='analyze_api')
 app.register_blueprint(env_api_blueprint, url_prefix='/qujata-api', name='env_info_api')
 
 CORS(app, origins=['*'])
